+++ imod/prepare/__init__.py
@@ -43,6 +43,8 @@ from imod.prepare.spatial import (
 from imod.prepare.topsystem import (
     ALLOCATION_OPTION,
     DISTRIBUTING_OPTION,
+    SimulationAllocationOptions,
+    SimulationDistributingOptions,
     allocate_drn_cells,
     allocate_ghb_cells,
     allocate_rch_cells,