+++ imod/tests/conftest.py
@@ -22,6 +22,10 @@ from .fixtures.flow_basic_unstructured_fixture import (
 )
 from .fixtures.flow_example_fixture import imodflow_model
 from .fixtures.flow_transport_simulation_fixture import flow_transport_simulation
+from .fixtures.imod5_cap_data import (
+    cap_data_sprinkling_grid,
+    cap_data_sprinkling_points,
+)
 from .fixtures.imod5_well_data import (
     well_duplication_import_prj,
     well_mixed_ipfs,
@@ -88,4 +92,6 @@ from .fixtures.mf6_welltest_fixture import (
     well_test_data_transient,
 )
 from .fixtures.msw_fixture import fixed_format_parser, simple_2d_grid_with_subunits
+from .fixtures.msw_imod5_cap_fixture import imod5_cap_data
+from .fixtures.msw_meteo_fixture import meteo_grids
 from .fixtures.msw_model_fixture import coupled_mf6_model, coupled_mf6wel, msw_model