+++ imod/typing/__init__.py
@@ -2,7 +2,7 @@
 Module to define type aliases.
 """
 
-from typing import TYPE_CHECKING, TypeAlias, TypeVar, Union
+from typing import TYPE_CHECKING, Literal, TypeAlias, TypedDict, TypeVar, Union
 
 import numpy as np
 import xarray as xr
@@ -11,11 +11,23 @@ from numpy.typing import NDArray
 
 GridDataArray: TypeAlias = Union[xr.DataArray, xu.UgridDataArray]
 GridDataset: TypeAlias = Union[xr.Dataset, xu.UgridDataset]
+GridDataDict: TypeAlias = dict[str, GridDataArray]
 ScalarAsDataArray: TypeAlias = Union[xr.DataArray, xu.UgridDataArray]
 ScalarAsDataset: TypeAlias = Union[xr.Dataset, xu.UgridDataset]
 UnstructuredData: TypeAlias = Union[xu.UgridDataset, xu.UgridDataArray]
 FloatArray: TypeAlias = NDArray[np.floating]
 IntArray: TypeAlias = NDArray[np.int_]
+
+
+class SelSettingsType(TypedDict, total=False):
+    layer: int
+    drop: bool
+    missing_dims: Literal["raise", "warn", "ignore"]
+
+
+class Imod5DataDict(TypedDict, total=False):
+    cap: GridDataDict
+    extra: dict[str, list[str]]
 
 
 # Types for optional dependencies.